refactor: move head util to `runtime/` directory
